move MaintenanceService to package
